+++ ouroboros-network/ouroboros-network.cabal
@@ -54,6 +54,7 @@ library
                        Ouroboros.Network.Channel
                        Ouroboros.Network.DeltaQ
                        Ouroboros.Network.Magic
+                       Ouroboros.Network.ErrorPolicy
                        Ouroboros.Network.Mux
                        Ouroboros.Network.NodeToNode
                        Ouroboros.Network.NodeToClient
@@ -180,6 +181,7 @@ test-suite test-network
                        Ouroboros.Network.Channel
                        Ouroboros.Network.DeltaQ
                        Ouroboros.Network.Magic
+                       Ouroboros.Network.ErrorPolicy
                        Ouroboros.Network.Mux
                        Ouroboros.Network.MockNode
                        Ouroboros.Network.NodeToNode
