+++ ouroboros-network/src/Ouroboros/Network/NodeToNode.hs
@@ -68,6 +68,7 @@ import           Network.TypedProtocol.Driver.ByteLimit (DecoderFailureOrTooMuch
 import           Network.TypedProtocol.Driver (TraceSendRecv)
 
 import           Ouroboros.Network.Magic
+import           Ouroboros.Network.ErrorPolicy
 import           Ouroboros.Network.Mux
 import           Ouroboros.Network.Protocol.Handshake.Type
 import           Ouroboros.Network.Protocol.Handshake.Version
@@ -86,7 +87,6 @@ import           Ouroboros.Network.Subscription.Dns ( DnsSubscriptionTarget (..)
                                                     , WithDomainName (..)
                                                     )
 import           Ouroboros.Network.Subscription.Worker (LocalAddresses (..))
-import           Ouroboros.Network.Subscription.PeerState
 
 
 -- | An index type used with the mux to enumerate all the mini-protocols that
