+++ ouroboros-network/src/Ouroboros/Network/Subscription/Worker.hs
@@ -55,6 +55,7 @@ import           Control.Monad.Class.MonadTimer
 import           Control.Monad.Class.MonadThrow
 import           Control.Tracer
 
+import           Ouroboros.Network.ErrorPolicy (CompleteApplication, Result (..))
 import           Ouroboros.Network.Server.ConnectionTable
 import           Ouroboros.Network.Subscription.Subscriber
 
@@ -78,36 +79,6 @@ maxConnectionAttemptDelay = 2 -- 2s delay
 --
 ipRetryDelay :: DiffTime
 ipRetryDelay = 10 -- 10s delay
-
-
--- | Result of the connection thread.  It's either result of an application, or
--- an exception thrown by it.
---
-data Result addr r where
-     ApplicationResult
-       :: !Time
-       -> !addr
-       -> !r
-       -> Result addr r
-
-     Connected
-       :: !Time
-       -> !addr
-       -> Result addr r
-
-     ConnectionError
-       :: Exception e
-       => !Time
-       -> !addr
-       -> !e
-       -> Result addr r
-
-     ApplicationError
-       :: Exception e
-       => !Time
-       -> !addr
-       -> !e
-       -> Result addr r
 
 data ResOrAct m addr r =
      Res !(Result addr r)
@@ -146,10 +117,6 @@ data SocketState m addr
 -- | Callback which fires: when we create or close a socket.
 --
 type SocketStateChange m s addr = SocketState m addr -> s -> STM m s
-
--- | Complete a connection, which receive application result (or exception).
---
-type CompleteApplication m s addr r = Result addr r -> s -> STM m (s, m ())
 
 -- | Given current state 'retry' too keep the subscription worker going.
 -- When this transaction returns, all the threads spawned by the worker will be
