+++ ouroboros-network/test/Test/PeerState.hs
@@ -32,6 +32,7 @@ import           Control.Monad.Class.MonadThrow
 import           Control.Tracer
 
 import           Data.Semigroup.Action
+import           Ouroboros.Network.ErrorPolicy
 import           Ouroboros.Network.Server.ConnectionTable
 import           Ouroboros.Network.Subscription.Ip
 import           Ouroboros.Network.Subscription.PeerState
